# rewriter optimization; switch things away from compareInternal