```
       django_template.py             4.6s (2)             4.4s (2)  -4.3%
            pyxl_bench.py             4.0s (2)             3.9s (2)  -1.6%
sqlalchemy_imperative2.py             5.3s (2)             5.2s (2)  -1.3%
        django_migrate.py             2.0s (2)             1.9s (2)  -5.4%
      virtualenv_bench.py             5.3s (2)             5.2s (2)  -1.2%
                  geomean                 4.0s                 3.9s  -2.8%
```